requirements.txt
pandas==1.3.5
Cython==0.29.28
scikit-learn==1.0.2
seaborn==0.11.2
tensorflow-gpu==2.8.0
tensorflow-probability==0.12.1
tensorflow-datasets==4.5.0
tf-models-official==2.6.1
numba==0.55.1
networkx==2.5.1
pydot==1.4.2
wandb==0.12.10
jupyter==1.0.0
jupyterlab==3.2.9
plotly==5.5.0
statsmodels
tensorboard-plugin-profile==2.5.0
